Refactor for ProviderAwareFile move
