+++ libs/create-deepagent/src/registry/files/model.ts
@@ -1,4 +1,4 @@
-import type { ProviderAwareFile } from "../framework.js";
+import type { ProviderAwareFile } from "../provider.js";
 
 /**
  * Create a ProviderAwareFile for `model.ts` that exports `coordinatorModel` and
