+++ libs/create-deepagent/src/registry/index.ts
@@ -3,10 +3,7 @@ export {
   type ProviderConfig,
   type EnvVarSpec,
 } from "./provider.js";
-export {
-  createFramework,
-  type FrameworkConfig,
-} from "./framework.js";
+export { createFramework, type FrameworkConfig } from "./framework.js";
 
 import { openai } from "../../registry/providers/openai.js";
 import { anthropic } from "../../registry/providers/anthropic.js";
